Rename user defined subscriber file
