+++ dds/src/implementation/dds_impl/domain_participant_impl.rs
@@ -72,7 +72,7 @@ use super::{
     builtin_subscriber_impl::BuiltInSubscriberImpl, message_receiver::MessageReceiver,
     participant_discovery::ParticipantDiscovery, status_condition_impl::StatusConditionImpl,
     status_listener::StatusListener, topic_impl::TopicImpl,
-    user_defined_publisher::UserDefinedPublisher, user_defined_subscriber::UserDefinedSubscriberImpl,
+    user_defined_publisher::UserDefinedPublisher, user_defined_subscriber_impl::UserDefinedSubscriberImpl,
 };
 
 pub const ENTITYID_SPDP_BUILTIN_PARTICIPANT_WRITER: EntityId =
