+++ dds/src/implementation/dds_impl/mod.rs
@@ -18,5 +18,5 @@ pub mod topic_impl;
 pub mod user_defined_data_reader;
 pub mod user_defined_data_writer;
 pub mod user_defined_publisher;
-pub mod user_defined_subscriber;
+pub mod user_defined_subscriber_impl;
 pub mod writer_factory;
