+++ dds/src/implementation/dds_impl/user_defined_data_reader.rs
@@ -58,7 +58,7 @@ use super::{
     any_data_reader_listener::AnyDataReaderListener, domain_participant_impl::AnnounceKind,
     message_receiver::MessageReceiver, status_condition_impl::StatusConditionImpl,
     status_listener::StatusListener, topic_impl::TopicImpl,
-    user_defined_subscriber::UserDefinedSubscriberImpl,
+    user_defined_subscriber_impl::UserDefinedSubscriberImpl,
 };
 
 pub enum UserDefinedReaderDataSubmessageReceivedResult {
